Bug fix
Clock no longer pauses before moving on to the next interval
Clock no longer pauses between empty intervals
This problem was due to the clock counting the zeroth second as its own
time unit which, for example caused a countdown from 1 to take two seconds
as it processed the zero as well. The same problem occurred when iterating
through empty intervals. Now the clock stops at 1, which caused the clock
to never count down to zero. This was fixed by hard coding a zero at the end
of the countdown.