New feature: restart timer
- Added a checkbox to the HTML which lets the user automatically restart the
  timer at the topmost interval
- Removed top inteval indicator in HTML
- Timer now highlights the current interval

Intervals were difficult to tell apart. Especially when all the values had
short, similar values it was easy to lose track of the current interval. The
restart button made this problem even more prevalent. A timer with intervals
2, 3, 2, 3 second intervals on repeat, for example, would count down as
expected, however there was no way to tell which particular interval was
currently counting down. I fixed the problem by highlighting the current
interval counting down.

Also, I added code to restart the loop when done if the user has opted to. This
was done with a call to startClock() when the clock has finished counting down.
This call is recursive in that startClock() again calls startCountdown(), but
this startCountdown() will return to startClock() before it calls itself
in setTimeout(). Then it will return  back to the first startCountdown(),
which will return and end the recursivity. In other words, this shouldn't
recurse forever.

startClock() now advances to the next available interval before the first call
to startCountdown(), and when none is available, returns. This is important as
it avoids infinite recursion for the case where the clock is completely empty
but is set to "Restart when done". In that case, the call stack would grow
infinitely as startClock() would call startCountdown(), which would then call
startClock() to restart the timer because it reached the last interval, which
would call startCountdown(), and so on.